+++ packages/charts/src/components/base-chart/base-chart.tsx
@@ -74,6 +74,19 @@ export type BaseChartProps = ResponsiveContainerProps & {
    */
   xDomain?: XAxisProps['domain'];
   /**
+   * Set the label formatter for the X axis
+   */
+  xFormatter?: XAxisProps['tickFormatter'];
+  /**
+   * Set the domain for the X axis.
+   */
+  xType?: XAxisProps['type'];
+  /**
+   * Allow the axis has duplicated categorys or not when the type of axis is "category".
+   * @url https://recharts.org/en-US/api/XAxis#allowDuplicatedCategory
+   */
+  xAllowDuplicatedCategory?: XAxisProps['allowDuplicatedCategory'];
+  /**
    * Set the domain for the left Y axis.
    */
   yDomainLeft?: YAxisProps['domain'];
@@ -81,10 +94,6 @@ export type BaseChartProps = ResponsiveContainerProps & {
    * Set the domain for the right Y axis.
    */
   yDomainRight?: YAxisProps['domain'];
-  /**
-   * Set the domain for the X axis.
-   */
-  xType?: XAxisProps['type'];
   /**
    * Set the domain for the left Y axis.
    */
@@ -181,7 +190,9 @@ export const BaseChart = forwardRef<HTMLDivElement, PropsClassChildren<BaseChart
   xType,
   dataKeyX = 'x',
   xPadding = 0,
+  xFormatter,
   xDomain,
+  xAllowDuplicatedCategory,
   density = 'mid',
   tooltipColors,
   customTooltip,
@@ -238,6 +249,7 @@ export const BaseChart = forwardRef<HTMLDivElement, PropsClassChildren<BaseChart
               type={xType}
               minTickGap={32}
               tick={{ fill: 'var(--dimmed-3)', fontSize: '0.8em' }}
+              tickFormatter={xFormatter}
               padding={{ left: xPadding, right: xPadding }}
               tickLine={{ stroke: 'var(--dimmed-3)' }}
               axisLine={{ stroke: 'var(--dimmed-4)' }}
@@ -245,6 +257,7 @@ export const BaseChart = forwardRef<HTMLDivElement, PropsClassChildren<BaseChart
               allowDataOverflow
               tickMargin={8}
               domain={xDomain}
+              allowDuplicatedCategory={xAllowDuplicatedCategory}
             />
             {referenceComponent}