+++ packages/charts/src/components/line-chart/line-chart.stories.tsx
@@ -2,6 +2,7 @@ import { Chip } from '@lualtek/react-components';
 import type { Meta, StoryObj } from '@storybook/react';
 
 import SimpleData from '../../../fixtures/data';
+import { MultiDataSeries } from '../../../fixtures/multi-data-axes';
 import MultiAxisData from '../../../fixtures/multi-y-data';
 import { ReferenceArea, ReferenceLine } from '../base-chart';
 import { ChartDataBaseType } from '../base-chart/base-chart';
@@ -217,5 +218,22 @@ export const WithReferenceAreasAndLines = {
 export const WithCustomYDomain = {
   args: {
     yDomainLeft: [0, 1000],
+  },
+} satisfies Story;
+
+export const WithFormattedXLabels = {
+  args: {
+    xFormatter: value => `${value} Formatted`,
+  },
+} satisfies Story;
+
+export const WithMultiDataAxes = {
+  args: {
+    data: undefined,
+    series: MultiDataSeries,
+    dataKeyX: 'category',
+    xType: 'category',
+    xAllowDuplicatedCategory: false,
+    yDomainLeft: [0, 1],
   },
 } satisfies Story;