+++ packages/react-components/src/components/selection-controls/checkbox/checkbox.tsx
@@ -2,7 +2,7 @@
 
 import { TokensTypes } from '@lualtek/tokens/platforms/web';
 import clsx from 'clsx';
-import { domAnimation, LazyMotion, m } from 'framer-motion';
+import { domAnimation, LazyMotion, m } from 'motion/react';
 import {
   ChangeEvent, forwardRef, InputHTMLAttributes, ReactNode, useCallback, useEffect, useId, useRef,
   useState,
@@ -159,7 +159,6 @@ export const Checkbox = forwardRef<HTMLInputElement, CheckboxProps>(({
         </Stack>
         {(invalid ?? isUserInvalid) && (
           <Stack
-            className={styles.Hint}
             hPadding={properties[dimension].text.padding}
           >
             <Text