+++ postcss.config.cjs
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
 const { postcss } = require('@lualtek/config');
 
 module.exports = postcss.getConfig();