+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @lualtek/icons
 
+## 14.3.3
+
+## 14.3.2
+
+### Patch Changes
+
+- 36d9dc0: Update deps
+
+## 14.3.1
+
+## 14.3.0
+
+## 14.2.0
+
+## 14.1.0
+
+## 14.0.3
+
+### Patch Changes
+
+- 4d4aa1f: Added optional data for LineChart series
+
 ## 14.0.2
 
 ### Patch Changes